requirements.txt
bs4==0.0.1
python-dateutil==2.8.2
requests==2.26.0
